[#9] refactor(naming)
