+++ src/main/kotlin/com/yapp/web2/domain/account/service/UserService.kt
@@ -1,7 +1,7 @@
-package com.yapp.web2.domain.user.service
+package com.yapp.web2.domain.account.service
 
-import com.yapp.web2.domain.user.entity.Account
-import com.yapp.web2.domain.user.repository.UserRepository
+import com.yapp.web2.domain.account.entity.Account
+import com.yapp.web2.domain.account.repository.UserRepository
 import com.yapp.web2.security.jwt.JwtProvider
 import com.yapp.web2.security.jwt.TokenDto
 import org.springframework.stereotype.Service
@@ -13,8 +13,7 @@ class UserService(
     private val jwtProvider: JwtProvider
 ) {
 
-    // email의 @ 이후를 제외하고 저장하는 게 좋아보이기도 한다. -> 맞는 거 같다
-    fun oauth2LoginUser(dto: Account.AccountRequest): TokenDto {
+    fun oauth2LoginUser(dto: Account.AccountLoginRequest): TokenDto {
         var account = Account.requestToAccount(dto)
         // 존재할 수도 있다
         account = userRepository.findByEmail(account.email)?.let {
